+++ docs/sources/CHANGELOG.md
@@ -15,6 +15,8 @@ The CHANGELOG for the current development version is available at
 ##### New Features
 
 - New `max_len` parameter for the frequent itemset generation via the `apriori` function to allow for early stopping. ([#270](https://github.com/rasbt/mlxtend/pull/270))
+- Added a `loadings_` attribute to `LinearDiscriminantAnalysis` to compute the factor loadings of the features on the components (discrimnants). ([#277](https://github.com/rasbt/mlxtend/pull/277))
+- Added a `tol` parameter to `LinearDiscriminantAnalysis` to threshold small eigenvalues, which are due to floating point imprecision, to zero for numerical stability. ([#277](https://github.com/rasbt/mlxtend/pull/277))
 
 ##### Changes
 
@@ -22,6 +24,9 @@ The CHANGELOG for the current development version is available at
 - The `SequentialFeatureSelector` now runs the continuation of the floating inclusion/exclusion as described in Novovicova & Kittler (1994). 
 Note that this didn't cause any difference in performance on any of the test scenarios but could lead to better performance in certain edge cases. 
 [#262](https://github.com/rasbt/mlxtend/pull/262)
+- Added "SVD solver" option to the `LinearDiscriminantAnalysis`. ([#277](https://github.com/rasbt/mlxtend/pull/277))
+- `LinearDiscriminantAnalysis` now uses NumPy's `cov` function instead of computing the scatter matrices manually to improve numerical stability. ([#277](https://github.com/rasbt/mlxtend/pull/277))
+
 
 
 ### Version 0.9.0 (2017-10-21)